chore(worker): make worker types more generic
BREAKING CHANGE: algorithm now expects to return `leftoverGenerations` instead of `excessGeneration`
