+++ packages/worker/src/index.ts
@@ -2,3 +2,4 @@ export * from './matching';
 export * from './matching-data';
 export * from './matching-result';
 export * from './builder';
+export * from './types';
