+++ packages/worker/src/matching-data/matching-data-in-memory-adapter/matching-data-in-memory.service.ts
@@ -1,5 +1,5 @@
 import { Injectable } from '@nestjs/common';
-import { MatchingInput } from '../types';
+import { MatchingInput } from '../../types';
 
 @Injectable()
 export class MatchingDataInMemoryService {
