+++ packages/worker/src/matching-data/types.ts
@@ -1,20 +1,3 @@
-export interface Reading {
-  deviceId: string;
-  volume: number;
-}
-
-export interface Preferences {
-  groupPriority: {
-    id: string;
-    groupPriority: { id: string }[][];
-  }[][];
-}
-
-export interface MatchingInput {
-  generations: Reading[];
-  consumptions: Reading[];
-  preferences: Preferences;
-  timestamp: Date;
-}
+import type { MatchingInput } from '../types';
 
 export type MatchCallback = (input: MatchingInput) => Promise<void>;
