+++ packages/worker/src/matching-result/index.ts
@@ -1,3 +1,2 @@
 export * from './matching-result.facade';
-export * from './types';
 export * from './matching-result-receivers-adapter';
