+++ packages/worker/src/matching-result/matching-result-receivers-adapter/types.ts
@@ -1,5 +1,4 @@
-import type { MatchingInput } from '../../matching-data';
-import type { MatchingResult } from '../types';
+import type { MatchingInput, MatchingResult } from '../../types';
 
 export type MatchingResultReceiver = (result: MatchingResult, input: MatchingInput) => Promise<void>;
 export const MATCHING_RESULT_RECEIVERS = Symbol.for('MATCHING_RESULT_RECEIVERS');
