--- packages/worker/src/matching-result/types.ts
@@ -1,12 +1,0 @@
-export interface MatchingResult {
-  timestamp: Date;
-  tree: {
-    rootHash: string;
-    leaves: string[];
-  },
-  data: {
-    matches: { consumerId: string; generatorId: string; volume: number }[];
-    leftoverConsumptions: { id: string; volume: number }[];
-    excessGenerations: { id: string; volume: number }[];
-  }
-}
